♻️ F08: Removed console logs and moved <EmptyMenu />
